requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anykeystore==0.2
# apex @ file:///data/taiye/Project/open-oasis/apex
async-timeout==5.0.1
attrs==24.2.0
av==13.1.0
bokeh==3.6.2
Brotli # @ file:///croot/brotli-split_1714483155106/work
cachetools # @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1724028158384/work
certifi # @ file:///croot/certifi_1725551672989/work/certifi
cffi==1.17.1
charset-normalizer # @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
cryptacular==1.6.2
decorator==5.1.1
decord==0.6.0
deepspeed==0.16.1
defusedxml==0.7.1
diffusers==0.31.0
docker-pycreds==0.4.0
einops==0.8.0
filelock # @ file:///croot/filelock_1700591183607/work
flash-attn==2.7.2.post1
frozenlist==1.5.0
fsspec==2024.10.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
gmpy2 # @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
greenlet==3.1.1
gym==0.26.2
gym-notices==0.0.8
hjson==3.1.0
huggingface-hub==0.26.2
hupper==1.12.1
hydra-core==1.3.2
idna # @ file:///croot/idna_1714398848350/work
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
internetarchive==5.0.5
iopath==0.1.10
Jinja2 # @ file:///croot/jinja2_1730902924303/work
jsonpatch==1.33
jsonpointer==3.0.0
lightning==2.4.0
lightning-utilities==0.11.9
MarkupSafe # @ file:///croot/markupsafe_1704205993651/work
mkl-service==2.4.0
mkl_fft # @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random # @ file:///io/mkl313/mkl_random_1730823916628/work
moviepy==2.1.1
mpmath # @ file:///croot/mpmath_1690848262763/work
msgpack==1.1.0
multidict==6.1.0
networkx # @ file:///croot/networkx_1720002482208/work
ninja==1.11.1.3
numpy # @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py # @ file:///home/conda/feedstock_root/build_artifacts/nvidia-ml-py_1698947663801/work
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop # @ file:///home/conda/feedstock_root/build_artifacts/nvitop_1726070205616/work
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
PasteDeploy==3.1.0
pbkdf2==1.3
pillow==10.4.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.3.6
plotly==5.24.1
portalocker==3.0.0
proglog==0.1.10
propcache==0.2.1
protobuf==5.29.1
psutil # @ file:///home/conda/feedstock_root/build_artifacts/psutil_1729847040822/work
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyramid==2.0.2
pyramid-mailer==0.15.1
PySocks # @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytorch-lightning==2.4.0
pytorchvideo==0.1.5
pytz==2024.2
PyYAML # @ file:///croot/pyyaml_1728657952215/work
rdkit==2024.3.6
regex==2024.11.6
repoze.sendmail==4.4.1
requests # @ file:///croot/requests_1730999120400/work
requests-oauthlib==2.0.0
rotary-embedding-torch==0.8.6
safetensors==0.4.5
scipy==1.14.1
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.1
soundfile==0.12.1
SQLAlchemy==2.0.36
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
termcolor # @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1728288976871/work
timm==1.0.11
# torch==2.5.1
# torchaudio==2.4.0
# torchmetrics==1.6.0
# torchvision==0.19.0
tornado==6.4.2
tqdm==4.67.1
transaction==5.0
translationstring==1.4
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3 # @ file:///croot/urllib3_1727769808118/work
velruse==1.1.1
venusian==3.1.1
wandb==0.19.0
wandb_osh==1.2.2
WebOb==1.8.9
wget==3.2
WTForms==3.2.1
wtforms-recaptcha==0.3.2
xyzservices==2024.9.0
yacs==0.1.8
yarl==1.18.3
zipp==3.21.0
zope.deprecation==5.0
zope.interface==7.2
zope.sqlalchemy==3.1
torchmetrics==1.6.1

